Load map from file, out of bounds workaround in player and ghost moviments and fix wrong direction UP and DOWN (they're inverted)

+++ src/com/pacman/basic/PrestigiousGhost.java
@@ -29,47 +29,50 @@ public class PrestigiousGhost extends Ghost{
 
 	@Override
 	public void move(LabyrinthMap map) {
-		//direção aleatória
-		Random r = new Random();
-		int x, y;
-		x = r.nextInt(map.getWidht());
-		y = r.nextInt(map.getHeight());
-
-		boolean foundWall = false;
-		boolean foundPlayer = false;
-		boolean foundGhost = false;
-
-		Player player = map.getPlayer();
-		Ghost ghosts[] = map.getGhosts();
-		Wall walls[] = map.getWall();
-		//verifica se nao é parede
-		for(int i = 0; i < map.getWall().length && !foundWall; i++) {
-			if(walls[i].isSameCoordinates(x, y))
-				foundWall = true;
-		}
-		//verifica se encontrou com jogador
-		if(player.isSameCoordinates(super.getX(), super.getY() + 1))
-			foundPlayer = true;
+		if(!this.isDead()) {
+			//direï¿½ï¿½o aleatï¿½ria
+			Random r = new Random();
+			int x, y;
+			x = r.nextInt(map.getWidht());
+			y = r.nextInt(map.getHeight());
 		
-		//verifica se não é fantasma
-		for(int i = 0; i < ghosts.length && !foundGhost; i++) {
-			if(ghosts[i].isSameCoordinates(super.getX(), super.getY() + 1))
-				foundGhost = true;
-		}
-		if(foundPlayer) {
-			//jogador foi pego por um fantasma random
-			if(player.isSuper()) {
-				//fantasma morre
-				this.killGhost();
+			boolean foundWall = false;
+			boolean foundPlayer = false;
+			boolean foundGhost = false;
+		
+			Player player = map.getPlayer();
+			Ghost ghosts[] = map.getGhosts();
+			Wall walls[] = map.getWall();
+			//Esse fantasma nunca estÃ¡ fora do limite do mapa devido ao calculo de x e y random ser de 0 -> limites;
+			//verifica se nao ï¿½ parede
+			for(int i = 0; i < map.getWall().length && !foundWall; i++) {
+				if(walls[i].isSameCoordinates(x, y))
+					foundWall = true;
 			}
-			else {
-				// jogador perde vida
-				player.kill();
+			//verifica se encontrou com jogador
+			if(player.isSameCoordinates(super.getX() + x, super.getY() + y))
+				foundPlayer = true;
+			
+			//verifica se nï¿½o ï¿½ fantasma
+			for(int i = 0; i < ghosts.length && !foundGhost; i++) {
+				if(ghosts[i].isSameCoordinates(super.getX() + x, super.getY() + y))
+					foundGhost = true;
 			}
-		} else if(!foundWall && !foundGhost) { //nao é parede e nem fantasma portanto pode teletransportar
-			super.getCoordinate().changeCoordinates(x, y);
-		} else {
-			this.move(map);
+			if(foundPlayer) {
+				//jogador foi pego por um fantasma random
+				if(player.isSuper()) {
+					//fantasma morre
+					this.killGhost();
+				}
+				else {
+					// jogador perde vida
+					player.kill();
+				}
+			} else if(!foundWall && !foundGhost) { //nao ï¿½ parede e nem fantasma portanto pode teletransportar
+				super.getCoordinate().changeCoordinates(x, y);
+			} else {
+				this.move(map);
+			}
 		}
 	}
